third_party/blink/web_tests/external/wpt/css/css-color/parsing/alpha-color-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property color value 'alpha(from red / none)'
  assert_equals: expected "color(srgb 1 0 0 / none)" but got "rgba(255, 0, 0, 0)"
Harness: the test ran to completion.
